requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
httpx==0.27.0
PyYAML==6.0.2
Jinja2==3.1.4
